% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantone.R
\docType{data}
\name{pantone}
\alias{pantone}
\title{pantone}
\format{
A data frame with 1341 rows and 2 variables:
\describe{
\item{pantone}{character The Pantone color name}
\item{hex}{character The associated hex value}
}
}
\source{
\url{"https://www.easycalculation.com/colorconverter/pantone-to-hex-tabl.php"}
}
\usage{
pantone
}
\description{
a database with hex values associated with 1341 Pantone colors
}
\keyword{datasets}
